Fix build failure on x86_64, due to MongoDB signature
Due to incorrect type conversion on x86_64 for MongoDB API, Vermont
fails to build. While it should be properly fixed in MongoDB C++ driver,
using a more generic static_cast here will solve it for now.
For more information, see https://jira.mongodb.org/browse/SERVER-4536

+++ src/modules/ipfix/IpfixDbWriterMongo.cpp
@@ -328,8 +328,8 @@ mongo::BSONObj IpfixDbWriterMongo::getInsertObj(const IpfixRecord::SourceID& sou
 		}
 
 		msg(MSG_DEBUG, "saw ipfix id %s in packet with intdata %llX", prop->propertyName,
-        static_cast<int64_t>(intdata));
-		obj << prop->propertyName << static_cast<int64_t>(intdata);
+        static_cast<long long int>(intdata));
+		obj << prop->propertyName << static_cast<long long int>(intdata);
 	}
 
 	if (flowstartsec == 0) {
